requirements.txt
matplotlib
pdf2image
pandas
streamlit
pytesseract
pdfplumber
docx2txt
PyPDF2
pyenchant